backend/requirements.txt
# Django and REST Framework
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0
django-ratelimit>=4.1,<5.0

# Database
psycopg2-binary>=2.9,<3.0

# Background Tasks
celery>=5.3,<6.0
redis>=4.5,<5.0
django-redis>=5.3,<6.0
django-celery-beat>=2.5,<3.0

# Web Scraping
playwright>=1.40,<2.0

# Environment and Configuration
python-decouple>=3.8,<4.0
dj-database-url>=2.1,<3.0

# API Documentation
drf-spectacular>=0.26,<1.0

# Development and Testing
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
pytest-asyncio>=0.21,<1.0
factory-boy>=3.3,<4.0

# Utilities
requests>=2.31,<3.0
python-dateutil>=2.8,<3.0
pillow>=10.0,<11.0

# AI/LLM Integration
openai>=1.3,<2.0

# Production
gunicorn>=21.2,<22.0
whitenoise>=6.5,<7.0